perf: 优化多进程的数据交换、事件生成和发送

+++ growingio-autotracker/src/main/java/com/growingio/android/sdk/autotrack/AutotrackConfiguration.java
@@ -70,13 +70,13 @@ public class AutotrackConfiguration extends TrackConfiguration {
     }
 
     @Override
-    public AutotrackConfiguration setDataUploadInterval(long dataUploadInterval) {
+    public AutotrackConfiguration setDataUploadInterval(int dataUploadInterval) {
         super.setDataUploadInterval(dataUploadInterval);
         return this;
     }
 
     @Override
-    public AutotrackConfiguration setSessionInterval(long sessionInterval) {
+    public AutotrackConfiguration setSessionInterval(int sessionInterval) {
         super.setSessionInterval(sessionInterval);
         return this;
     }
